🐛 Fix edit use not working
After the database migration on release 3.0.0, the insert stopped
behaving like an Upsert, as it used to in Objectbox.

This commit changes the insert in the database to be an Upsert, as
specified in https://www.sqlite.org/lang_UPSERT.html. The changes were
made to file Use.sq.

This behavior is only to the database, and thus a single test was added
to UseRepositoryTest to ensure it. All other files changed are
collateral of the rename refactor.

Closes #124

Signed-off-by: Leonardo Colman Lopes <dev@leonardo.colman.com.br>