✨ Refactor SymptomsPage for more withdrawal symptoms
Replaced the ThcConcentrationView and DiscomfortView with a direct use
of the UseRepository on the SymptomsPage. This refactor makes the
architecture simpler and code more readable. Further, added new data
related to withdrawal symptoms and displayed them using the new
WithdrawalChart view. Also, updated the
WidgetConcentrationDiscomfortPart and added the necessary strings to the
 resources file.

Signed-off-by: Leonardo Colman Lopes <dev@leonardo.colman.com.br>