:sparkles: Add a Day/Night switch that doesn't depend on the System Theme
This commit introduces a new feature in the settings repository where
users can now toggle between dark and light mode. The changes
include the addition of a boolean key 'darkMode' in the settings
repository. Also, functions to set and retrieve the state of the
'darkMode' have been introduced. New test cases have been added to
ensure the corret behavior of the 'darkMode'.