🛂 Replace dependency license check with Licensee plugin
The dependency license check provided by the jk1 plugin has been
replaced with Licensee plugin to manage the allowed licenses. This
modification includes changing the Gradle and GitHub workflow
configurations to use the Licensee plugin for license checks. The list
of allowed licenses is now defined directly in the app's build.gradle
.kts.

Signed-off-by: Leonardo Colman Lopes <dev@leonardo.colman.com.br>